pyproject.toml
[tool.poetry]
name = "python-service-project"
version = "0.1.0"
description = ""
authors = ["Egor Nikitin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
click = "^8.1.7"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
google-api-python-client = "^2.143.0"
google-auth = "^2.34.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
vulture = "^2.11"


[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    )/

)
'''


[tool.isort]
line_length = 119


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
